Improved colour selection algo some more at the cost of performance: - Re-iterative search implemented (with a hard limit on the number of retry-attempts allowed) which adjusts the sample size when too many distinct colours were picked, in an attempt to reduce the size of the output until it fits the colour grid. - Changed distance to a decimal input with 2 decimal places. It's currently quite hard to select exact figures but this will be easier once the dual input system for selecting it is implemented.

+++ main.js
@@ -29,7 +29,7 @@ function updateRootColour(colour) {
 
 function updateDistance(d) {
     // TODO: validation
-    distance = parseInt(d);
+    distance = parseFloat(d);
 }
 
 function updateRootColourDisplay() {
